Planner was really addprogram since the beginning. Renamed and fixed routes
