+++ frontend/src/components/App.tsx
@@ -3,9 +3,11 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 import Header from './Header'
 import Homepage from './Homepage'
 import Planner from './Planner'
+import AddProgram from './AddProgram'
 import Training from './Training'
 import Workouts from './Workouts'
 import GlobalStyle from '../styles/GlobalStyles'
+
 const App: React.FC = () => {
   return (
     <Router>
@@ -13,10 +15,11 @@ const App: React.FC = () => {
       <Header />
       <Routes>
         <Route path="/" element={<Homepage />} />
-        <Route path="/planner" element={<Planner />} />
         <Route path="/workouts" element={<Workouts />} />
-        {/* TODO: Add route /planner/programs & /planner/programs/:programId */}
         <Route path="/workouts/:workoutId" element={<Training />} />
+        {/* TODO: Add route /planner/programs & /planner/programs/:programId */}
+        <Route path="/planner" element={<Planner />} />
+        <Route path="/planner/addprogram" element={<AddProgram />} />
       </Routes>
     </Router>
   )
